test: move tests to correct locations
