+++ packages/esql-parser/src/esql/__tests__/parens.test.ts
@@ -5,13 +5,11 @@
  * 2.0.
  */
 
-import { Parser } from '..';
+import { Parser } from '../parser';
 import type { ESQLParens } from '@elastic/esql-types';
-import { BasicPrettyPrinter } from '../../pretty_print';
 
 const parse = (src: string) => Parser.parse(src, { withParens: true }).root;
 const firstArg = (src: string) => parse(src).commands[0].args[0];
-const reprint = (src: string) => BasicPrettyPrinter.print(parse(src));
 
 describe('expression parens are dropped by default', () => {
   const firstArgDefault = (src: string) => Parser.parse(src).root.commands[0].args[0];
@@ -164,38 +162,6 @@ describe('expression parens preserved with { withParens: true }', () => {
       const expr = firstArg('ROW (a + b)') as ESQLParens;
 
       expect(expr.text).toBe('(a+b)');
-    });
-  });
-
-  describe('round-trip: printer preserves source parens', () => {
-    it('redundant same-precedence right-operand parens are preserved', () => {
-      expect(reprint('FROM a | WHERE a + (b + c)')).toBe('FROM a | WHERE a + (b + c)');
-      expect(reprint('FROM a | WHERE a * (b * c)')).toBe('FROM a | WHERE a * (b * c)');
-      expect(reprint('FROM a | WHERE a - (b + c)')).toBe('FROM a | WHERE a - (b + c)');
-    });
-
-    it('redundant left-operand parens are preserved', () => {
-      expect(reprint('FROM a | WHERE (b / c) * 10')).toBe('FROM a | WHERE (b / c) * 10');
-      expect(reprint('FROM a | WHERE (b + c) AND d')).toBe('FROM a | WHERE (b + c) AND d');
-    });
-
-    it('semantically significant parens are preserved', () => {
-      expect(reprint('FROM a | WHERE (a + b) * c')).toBe('FROM a | WHERE (a + b) * c');
-      expect(reprint('FROM a | WHERE a / (b * c)')).toBe('FROM a | WHERE a / (b * c)');
-      expect(reprint('FROM a | WHERE b AND (c OR d)')).toBe('FROM a | WHERE b AND (c OR d)');
-    });
-
-    it('nested parens are preserved', () => {
-      expect(reprint('FROM a | WHERE (1 + (2 + 3))')).toBe('FROM a | WHERE (1 + (2 + 3))');
-    });
-
-    it('NOT (expr) parens are preserved', () => {
-      expect(reprint('FROM a | WHERE NOT (a OR b)')).toBe('FROM a | WHERE NOT (a OR b)');
-      expect(reprint('FROM a | WHERE NOT (a > b)')).toBe('FROM a | WHERE NOT (a > b)');
-    });
-
-    it('inline cast with parenthesized inner expression', () => {
-      expect(reprint('ROW (1 + 2)::string')).toBe('ROW (1 + 2)::STRING');
     });
   });
